ui.relative_layout: fix incorrect type casting
Summary: `temph` can be double type.

Reviewers: Jaehyun_Cho

Reviewed By: Jaehyun_Cho

Subscribers: cedric, #reviewers, #committers

Tags: #efl

Differential Revision: https://phab.enlightenment.org/D8695

+++ src/lib/elementary/efl_ui_relative_layout.c
@@ -125,7 +125,7 @@ static void
 _child_aspect_calc(Efl_Ui_Relative_Layout_Child *child, Eina_Bool axis)
 {
    Efl_Ui_Relative_Layout_Calc *calc = &child->calc;
-   int temph;
+   double temph;
 
    if ((calc->aspect[0] <= 0) || (calc->aspect[1] <= 0))
      {
